support/image-hosting/heihost/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "heihost"
version = "0.1.0"
description = "Hassio E-Ink Image Server"
authors = ["Dennis Sitelew <dsitelew@yobasoft.de>"]
packages = [
    { include = "heihost", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.4.0"
numpy = "^2.1.1"
selenium = "^4.24.0"
schedule = "^1.2.2"
lz4 = "^4.3.3"

[tool.poetry.scripts]
hei-capture = "heihost.image_capture:sync_main"
hei-server = "heihost.server:sync_main"
